Add files via upload
The project activities involves the installation of 5 independent biogas digesters beginning 30/04/2016. The purpose of the project activity is the set up independent biogas plants (digesters) between the 600 m3 and 1600 m3 capacity range, for serving the captive electricity needs at the location of the project activity. Fresh poultry litter is fed into the anaerobic digesters.